A bunch of fixes and add AggregationMinutes

+++ ThinkOrSwimAlerts/Data/Models/PositionUpdate.cs
@@ -13,7 +13,7 @@ namespace ThinkOrSwimAlerts.Data.Models
         [Key]
         [DatabaseGenerated( DatabaseGeneratedOption.Identity )]
         public long PositionUpdateId { get; set; }
-        public uint PositionId { get; set; }
+        public long PositionId { get; set; }
         public Position Position { get; set; }
         public int SecondsAfterFirstBuy { get; set; }
         public float Mark { get; set; }
